backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
sse-starlette>=2.1.0
httpx>=0.27.0
sentence-transformers>=3.0.0
qdrant-client>=1.12.0
pymupdf>=1.24.0
python-docx>=1.1.0
numpy>=1.26.0
websockets>=13.0
